Made changes to try to fix issue number two.
Changes made on file /Admin/adminAlumnosNuevo.html include removing
the optional fields' inputs, as well as changing the javascript to
setup event handlers a little bit.

Changes made on file /js/main.js include adding proper regular
expressions for emails and urls, as well as handling the onclick
event of the checkbox on the previous file, so that they appear when
the checkbox is checked, and are removed when it is not.